backend/requirements.txt
SQLAlchemy
python-dotenv
fastapi
uvicorn
psycopg2-binary
pydantic
httpx
requests
pydantic[email]
